requirements.txt
streamlit == 1.29.0
matplotlib == 3.8.2
seaborn == 0.13.1
pandas
numpy
textblob
googletrans == 3.1.0a0
nltk
typing
nlp_id
pillow
xlsxwriter